// src/models/enums/ServicesEnum.ts
export enum ServicesEnum {
	SESSION_SERVICE = "F2F_SESSION_SERVICE",
	SESSION_CONFIG_SERVICE = "F2F_SESSION_CONFIG_SERVICE",
	AUTHORIZATION_SERVICE = "F2F_AUTHORIZATION_SERVICE",
	ACCESS_TOKEN_SERVICE = "F2F_ACCESS_TOKEN_SERVICE",
	DOCUMENT_SELECTION_SERVICE = "F2F_DOCUMENT_SELECTION_SERVICE",
	JWKS_SERVICE = "F2F_JWKS_SERVICE",
	USERINFO_SERVICE = "F2F_USER_INFO_SERVICE",
	GOV_NOTIFY_SERVICE = "F2F_GOV_NOTIFY_SERVICE",
	CALLBACK_SERVICE = "F2F_CALLBACK_SERVICE",
	ABORT_SERVICE = "F2F_ABORT_SERVICE",
	REMINDER_SERVICE = "F2F_REMINDER_SERVICE",
	THANK_YOU_EMAIL_SERVICE = "F2F_THANK_YOU_EMAIL_SERVICE",
	PERSON_INFO_SERVICE = "PERSON_INFO_SERVICE",
	PUBLIC_KEY_SSM_PATH_SERVICE = "PUBLIC_KEY_SSM_PATH_SERVICE",
	NA = "NA",
}