// src/models/enums/AuthSessionState.ts
export enum AuthSessionState {
	F2F_SESSION_CREATED = "F2F_SESSION_CREATED",
	F2F_YOTI_SESSION_CREATED = "F2F_YOTI_SESSION_CREATED",
	F2F_AUTH_CODE_ISSUED = "F2F_AUTH_CODE_ISSUED",
	F2F_ACCESS_TOKEN_ISSUED = "F2F_ACCESS_TOKEN_ISSUED",
	F2F_YOTI_SESSION_COMPLETE = "F2F_YOTI_SESSION_COMPLETE",
	F2F_CREDENTIAL_ISSUED = "F2F_CREDENTIAL_ISSUED",
	F2F_SESSION_ABORTED = "F2F_SESSION_ABORTED",
	F2F_CRI_SESSION_ABORTED = "F2F_CRI_SESSION_ABORTED",
}