// src/models/enums/TxmaEvents.ts
export enum TxmaEventNames {
	F2F_CRI_START = "F2F_CRI_START",
	F2F_CRI_AUTH_CODE_ISSUED = "F2F_CRI_AUTH_CODE_ISSUED",
	F2F_YOTI_PDF_EMAILED = "F2F_YOTI_PDF_EMAILED",
	F2F_YOTI_PDF_LETTER_POSTED = "F2F_YOTI_PDF_LETTER_POSTED",
	F2F_YOTI_START = "F2F_YOTI_START",
	F2F_YOTI_RESPONSE_RECEIVED = "F2F_YOTI_RESPONSE_RECEIVED",
	F2F_CRI_VC_ISSUED = "F2F_CRI_VC_ISSUED",
	F2F_CRI_END = "F2F_CRI_END",
	F2F_DOCUMENT_UPLOADED = "F2F_DOCUMENT_UPLOADED",
	F2F_CRI_SESSION_ABORTED = "F2F_CRI_SESSION_ABORTED",
}